+++ workflow_engine/project_apps/api/views.py
@@ -4,7 +4,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 
-from project_apps.service.workflow_service import WorkflowService, WorkflowExecutor
+from project_apps.service.workflow_service import WorkflowService
 
 
 class WorkflowCreateAPIView(APIView):
@@ -102,8 +102,8 @@ class WorkflowExecuteAPIView(APIView):
         Request data
         - uuid: workflow's uuid that you want to execute.
         '''
-        workflow_executor = WorkflowExecutor()
-        result = workflow_executor.execute_workflow(workflow_uuid)
+        workflow_service = WorkflowService()
+        result = workflow_service.execute_workflow(workflow_uuid)
 
         if result:
             return Response(status=status.HTTP_204_NO_CONTENT)
