+++ workflow_engine/project_apps/engine/job_execute.py
@@ -4,36 +4,36 @@ from requests.exceptions import ReadTimeout, ConnectionError
 
 from project_apps.constants import JOB_STATUS_RUNNING, JOB_STATUS_SUCCESS, JOB_STATUS_FAIL
 from project_apps.repository.history_repository import HistoryRepository
-from project_apps.service.workflow_service import WorkflowExecutor
+from project_apps.service.workflow_manage import WorkflowManager
 
 
 @shared_task
 def job_execute(workflow_uuid, history_uuid, job_uuid):
     client = docker.from_env()
-    workflow_executor = WorkflowExecutor()
+    workflow_manager = WorkflowManager()
     history_repo = HistoryRepository()
 
-    job_data = workflow_executor.find_job_data(workflow_uuid, job_uuid)
+    job_data = workflow_manager.find_job_data(workflow_uuid, job_uuid)
     
     if not job_data:    
         return
 
     try:
-        workflow_executor.update_job_status(workflow_uuid, job_uuid, JOB_STATUS_RUNNING)
+        workflow_manager.update_job_status(workflow_uuid, job_uuid, JOB_STATUS_RUNNING)
         image = client.images.pull(job_data['image'])
         container = client.containers.run(image, detach=True)
         # container = client.containers.run(image, detach=True, environment=job_data.get('parameters', {}))
         result = container.wait(timeout=60)
         
         if result['StatusCode'] == 0:
-            workflow_executor.update_job_status(workflow_uuid, job_uuid, JOB_STATUS_SUCCESS)
-            workflow_executor.handle_success(job_data, workflow_uuid, history_uuid, history_repo)
+            workflow_manager.update_job_status(workflow_uuid, job_uuid, JOB_STATUS_SUCCESS)
+            workflow_manager.handle_success(job_data, workflow_uuid, history_uuid, history_repo)
         else:
-            workflow_executor.handle_failure(history_uuid, workflow_uuid, history_repo)
+            workflow_manager.handle_failure(history_uuid, workflow_uuid, history_repo)
             
     except (ReadTimeout, ConnectionError) as e:
-        workflow_executor.update_job_status(workflow_uuid, job_uuid, JOB_STATUS_FAIL)
-        workflow_executor.handle_failure(history_uuid, workflow_uuid, history_repo)
+        workflow_manager.update_job_status(workflow_uuid, job_uuid, JOB_STATUS_FAIL)
+        workflow_manager.handle_failure(history_uuid, workflow_uuid, history_repo)
 
     except Exception as e:
-        workflow_executor.handle_failure(history_uuid, workflow_uuid, history_repo)
+        workflow_manager.handle_failure(history_uuid, workflow_uuid, history_repo)
